cli/pyproject.toml
[project]
name = "haymaker-cli"
version = "0.1.0"
description = "Command-line interface for Azure HayMaker service management"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Azure HayMaker Team", email = "team@azurehaymaker.dev"}
]
keywords = ["azure", "cli", "haymaker", "orchestration", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "click>=8.1.7",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "azure-identity>=1.16.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
    "pyright>=1.1.0",
    "respx>=0.21.0",
]

[project.scripts]
haymaker = "haymaker_cli.main:cli"

[project.urls]
Homepage = "https://github.com/azurehaymaker/azurehaymaker"
Documentation = "https://github.com/azurehaymaker/azurehaymaker/tree/main/docs"
Repository = "https://github.com/azurehaymaker/azurehaymaker"
"Bug Tracker" = "https://github.com/azurehaymaker/azurehaymaker/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/haymaker_cli"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=haymaker_cli",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports
"tests/**" = ["S101"]     # assert allowed in tests

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__", "**/.venv", "**/node_modules"]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
